feat: Enhance agent token tracking and telemetry configuration
- Added per-agent token usage tracking in LiveOrchestrator to monitor input/output tokens and response counts.
- Implemented emission of invoke_agent summary spans when switching agents, ensuring accurate token attribution.
- Refactored telemetry configuration to simplify setup and improve maintainability, including environment variable handling.
- Introduced FilteringSpanProcessor to combine noisy span filtering and PII scrubbing.
- Updated logging configuration to suppress noisy loggers and improve log clarity.
- Removed deprecated code and improved overall structure for better readability and performance.
