+++ utils/telemetry_config.py
@@ -3,516 +3,413 @@
 # Licensed under the MIT License. See License in the project root for
 # license information.
 # --------------------------------------------------------------------------
+"""
+Azure Monitor / Application Insights telemetry configuration.
+
+This module provides a simplified, maintainable setup for OpenTelemetry with Azure Monitor.
+
+Configuration via environment variables:
+- APPLICATIONINSIGHTS_CONNECTION_STRING: Required for Azure Monitor export
+- DISABLE_CLOUD_TELEMETRY: Set to "true" to disable all cloud telemetry
+- AZURE_MONITOR_DISABLE_LIVE_METRICS: Disable live metrics stream
+- AZURE_TRACING_GEN_AI_CONTENT_RECORDING_ENABLED: Record GenAI prompts/completions
+- TELEMETRY_PII_SCRUBBING_ENABLED: Enable PII scrubbing (default: true)
+
+See utils/pii_filter.py for PII scrubbing configuration options.
+"""
+
+from __future__ import annotations
+
 import logging
 import os
 import re
+import socket
+import uuid
 import warnings
+from dataclasses import dataclass, field
+from typing import List, Optional, Pattern
 
-# Suppress OpenTelemetry deprecation warnings about LogRecord init
-# This warning was fixed in opentelemetry-instrumentation-openai-v2>=2.2b0
-# Keep suppression as fallback for users with older versions
-warnings.filterwarnings(
-    "ignore",
-    message="LogRecord init with.*is deprecated",
-    module="opentelemetry"
-)
+# Suppress OpenTelemetry deprecation warnings
+warnings.filterwarnings("ignore", message="LogRecord init with.*is deprecated", module="opentelemetry")
 
-# Ensure environment variables from .env are available BEFORE we check DISABLE_CLOUD_TELEMETRY.
-try:  # minimal, silent if python-dotenv missing
-    from dotenv import load_dotenv  # type: ignore
-
-    # Only load if it looks like a .env file exists and variables not already present
+# Load .env early
+try:
+    from dotenv import load_dotenv
     if os.path.isfile(".env"):
         load_dotenv(override=False)
 except Exception:
     pass
 
-from azure.core.exceptions import HttpResponseError, ServiceResponseError
-from utils.azure_auth import get_credential, ManagedIdentityCredential
-from azure.monitor.opentelemetry import configure_azure_monitor
-from opentelemetry.sdk.resources import Resource, ResourceAttributes
-from opentelemetry.trace import SpanKind
-
-# Set up logger for this module (needs to be before functions that use it)
 logger = logging.getLogger(__name__)
-_live_metrics_permanently_disabled = False
-_azure_monitor_configured = False
+
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# OPENAI CLIENT INSTRUMENTOR (opentelemetry-instrumentation-openai-v2)
-# Auto-instruments OpenAI Python client for GenAI semantic conventions
+# CONFIGURATION
 # ═══════════════════════════════════════════════════════════════════════════════
 
-_openai_instrumentor_enabled = False
-
-def _setup_openai_instrumentation(tracer_provider=None) -> bool:
-    """
-    Enable OpenAI client instrumentation for automatic tracing.
+@dataclass
+class TelemetryConfig:
+    """Centralized telemetry configuration loaded from environment."""
     
-    IMPORTANT: This MUST be called BEFORE creating any OpenAI/AzureOpenAI clients.
-    The instrumentor monkey-patches the openai module at import time.
+    # Core settings
+    enabled: bool = True
+    connection_string: Optional[str] = None
+    logger_name: str = "default"
     
-    This uses opentelemetry-instrumentation-openai-v2 which automatically creates
-    spans with GenAI semantic conventions for:
-    - chat.completions API calls (including AzureOpenAI)
-    - Token usage tracking (gen_ai.usage.input_tokens, gen_ai.usage.output_tokens)
-    - Model information (gen_ai.request.model)
-    - Streaming completions
+    # Service identification (for Application Map)
+    service_name: str = "rtagent-api"
+    service_namespace: str = "callcenter-app"
+    service_version: Optional[str] = None
+    environment: Optional[str] = None
     
-    Content recording (prompts/completions) can be enabled via:
-    - AZURE_TRACING_GEN_AI_CONTENT_RECORDING_ENABLED=true environment variable
+    # Feature flags
+    enable_live_metrics: bool = True
+    enable_pii_scrubbing: bool = True
     
-    Args:
-        tracer_provider: Optional TracerProvider. If None, uses the global one.
+    # Instrumentation options
+    # NOTE: configure_azure_monitor() enables many auto-instrumentations by default.
+    # We explicitly disable most to prevent duplicate spans alongside our manual instrumentation.
+    enabled_instrumentations: List[str] = field(default_factory=lambda: [
+        # Only enable specific instrumentations we need
+        # "azure_sdk",  # Can cause noisy spans
+    ])
+    disabled_instrumentations: List[str] = field(default_factory=lambda: [
+        # Disable auto-instrumentations that would create duplicate spans
+        # alongside our manual instrumentation
+        "fastapi",      # We create manual spans for endpoints
+        "asgi",         # We create manual spans for ASGI lifecycle
+        "aiohttp",      # We have manual HTTP tracing
+        "requests",     # We have manual HTTP tracing  
+        "urllib3",      # We have manual HTTP tracing
+        "urllib",       # We have manual HTTP tracing
+        "httpx",        # We have manual HTTP tracing
+        "redis",        # We create manual Redis spans where needed
+        "azure_sdk",    # Can be very noisy, disable by default
+        "psycopg2",     # Not used
+        "django",       # Not used
+        "flask",        # Not used
+    ])
     
-    Returns:
-        True if instrumentor was enabled successfully, False otherwise.
-    """
-    global _openai_instrumentor_enabled
-    
-    if _openai_instrumentor_enabled:
-        return True
-    
-    # Check if early-init module already instrumented OpenAI
-    try:
-        from utils.openai_instrumentation import is_instrumented
-        if is_instrumented():
-            _openai_instrumentor_enabled = True
-            logger.debug("OpenAI instrumentation already enabled via early-init module")
-            return True
-    except ImportError:
-        pass  # Early-init module not available
-    
-    try:
-        from opentelemetry.instrumentation.openai_v2 import OpenAIInstrumentor
-        from opentelemetry import trace as otel_trace
+    @classmethod
+    def from_env(cls) -> "TelemetryConfig":
+        """Create configuration from environment variables."""
+        def _bool_env(key: str, default: bool) -> bool:
+            return os.getenv(key, str(default)).lower() not in ("false", "0", "no")
         
-        # Check if openai module was already imported (late instrumentation warning)
-        import sys
-        if "openai" in sys.modules:
-            logger.warning(
-                "⚠️ OpenAI module was imported before instrumentation setup! "
-                "GenAI spans may not work correctly. Import utils.openai_instrumentation first."
-            )
-        
-        # Enable content recording if configured
-        # This captures gen_ai.request.messages and gen_ai.response.choices
-        content_recording = os.getenv(
-            "AZURE_TRACING_GEN_AI_CONTENT_RECORDING_ENABLED", "false"
-        ).lower() == "true"
-        
-        if content_recording:
-            # Ensure environment variable is set for SDK to pick up
-            os.environ["AZURE_TRACING_GEN_AI_CONTENT_RECORDING_ENABLED"] = "true"
-            logger.info("GenAI content recording enabled (prompts/completions will be traced)")
-        
-        # Use provided tracer_provider or get the global one
-        tp = tracer_provider or otel_trace.get_tracer_provider()
-        OpenAIInstrumentor().instrument(tracer_provider=tp)
-        
-        _openai_instrumentor_enabled = True
-        logger.info(
-            "✅ OpenAI client instrumentation enabled (openai-v2)",
-            extra={"content_recording": content_recording}
+        return cls(
+            enabled=not _bool_env("DISABLE_CLOUD_TELEMETRY", False),
+            connection_string=os.getenv("APPLICATIONINSIGHTS_CONNECTION_STRING"),
+            logger_name=os.getenv("AZURE_MONITOR_LOGGER_NAME", "default"),
+            service_name=os.getenv("SERVICE_NAME", "rtagent-api"),
+            service_namespace=os.getenv("SERVICE_NAMESPACE", "callcenter-app"),
+            service_version=os.getenv("SERVICE_VERSION") or os.getenv("APP_VERSION"),
+            environment=os.getenv("ENVIRONMENT"),
+            enable_live_metrics=not _bool_env("AZURE_MONITOR_DISABLE_LIVE_METRICS", False),
+            enable_pii_scrubbing=_bool_env("TELEMETRY_PII_SCRUBBING_ENABLED", True),
         )
-        return True
-        
-    except ImportError:
-        logger.debug(
-            "opentelemetry-instrumentation-openai-v2 not available - "
-            "install with: pip install opentelemetry-instrumentation-openai-v2"
-        )
-        return False
-    except Exception as e:
-        logger.warning(f"Failed to enable OpenAI client instrumentation: {e}")
-        return False
-
-
-def is_openai_instrumented() -> bool:
-    """Check if OpenAI client instrumentation is enabled."""
-    return _openai_instrumentor_enabled
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# NOISY SPAN FILTERING - Filter high-frequency spans at export time
+# SPAN FILTERING
 # ═══════════════════════════════════════════════════════════════════════════════
 
-# ═══════════════════════════════════════════════════════════════════════════════
-# SPAN FILTERING CONFIGURATION
-# Define patterns for spans to drop (used by NoisySpanFilterProcessor)
-# ═══════════════════════════════════════════════════════════════════════════════
-
-# Patterns for span names that should ALWAYS be dropped (pure noise)
-NOISY_SPAN_PATTERNS = [
+# Patterns for noisy spans to drop
+NOISY_SPAN_PATTERNS: List[Pattern[str]] = [
     re.compile(r".*websocket\s*(receive|send).*", re.IGNORECASE),
     re.compile(r".*ws[._](receive|send).*", re.IGNORECASE),
     re.compile(r"HTTP.*websocket.*", re.IGNORECASE),
     re.compile(r"^(GET|POST)\s+.*(websocket|/ws/).*", re.IGNORECASE),
-    # Internal frame/chunk processing (high frequency, low value)
-    re.compile(r".*audio[._]chunk.*", re.IGNORECASE),
-    re.compile(r".*audio[._]frame.*", re.IGNORECASE),
-    re.compile(r".*process[._]frame.*", re.IGNORECASE),
-    re.compile(r".*stream[._]chunk.*", re.IGNORECASE),
-    re.compile(r".*emit[._]chunk.*", re.IGNORECASE),
-    # Redis polling/connection maintenance
+    re.compile(r".*audio[._](chunk|frame).*", re.IGNORECASE),
+    re.compile(r".*(process|stream|emit)[._](frame|chunk).*", re.IGNORECASE),
     re.compile(r".*redis[._](ping|pool|connection).*", re.IGNORECASE),
-    # Session polling
-    re.compile(r".*poll[._]session.*", re.IGNORECASE),
-    re.compile(r".*session[._]heartbeat.*", re.IGNORECASE),
+    re.compile(r".*(poll|heartbeat)[._]session.*", re.IGNORECASE),
+    # VoiceLive high-frequency streaming events
+    re.compile(r"voicelive\.event\.response\.audio\.delta", re.IGNORECASE),
+    re.compile(r"voicelive\.event\.response\.audio_transcript\.delta", re.IGNORECASE),
+    re.compile(r"voicelive\.event\.response\.function_call_arguments\.delta", re.IGNORECASE),
+    re.compile(r"voicelive\.event\.response\.text\.delta", re.IGNORECASE),
+    re.compile(r"voicelive\.event\.response\.content_part\.delta", re.IGNORECASE),
+    re.compile(r"voicelive\.event\.input_audio_buffer\.", re.IGNORECASE),
 ]
 
-# URL patterns that indicate noisy operations (used by NoisySpanFilterProcessor)
-NOISY_URL_PATTERNS = [
-    "/api/v1/browser/conversation",
-    "/api/v1/acs/media", 
-    "/ws/",
-    "/api/v1/health",
-    "/api/v1/readiness",
+# Loggers to suppress (set to WARNING level)
+NOISY_LOGGERS = [
+    "azure.identity", "azure.core.pipeline", "azure.monitor.opentelemetry.exporter",
+    "websockets", "aiohttp", "httpx", "httpcore",
+    "uvicorn.protocols.websockets", "uvicorn.error", "uvicorn.access",
+    "starlette.routing", "fastapi",
+    "opentelemetry.sdk.trace", "opentelemetry.exporter",
+    "redis.asyncio.connection",
 ]
+
+
+def _suppress_noisy_loggers(level: int = logging.WARNING) -> None:
+    """Set noisy loggers to WARNING level to reduce noise."""
+    for name in NOISY_LOGGERS:
+        logging.getLogger(name).setLevel(level)
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# NOISY SPAN FILTER PROCESSOR
-# SpanProcessor that filters noisy spans at export time
-# This catches spans created by auto-instrumentation that bypass the sampler
+# SPAN PROCESSOR WITH FILTERING AND PII SCRUBBING  
 # ═══════════════════════════════════════════════════════════════════════════════
 
 from opentelemetry.sdk.trace import SpanProcessor, ReadableSpan
 
-class NoisySpanFilterProcessor(SpanProcessor):
+
+class FilteringSpanProcessor(SpanProcessor):
     """
-    SpanProcessor that filters noisy spans at export time.
+    SpanProcessor that filters noisy spans and scrubs PII from attributes.
     
-    Auto-instrumented spans (from FastAPI, ASGI, etc.) may bypass the sampler
-    if they inherit sampling decisions from parent spans. This processor
-    filters those spans before they are exported to Azure Monitor.
+    Combines noise filtering and PII scrubbing in a single processor
+    for better performance and simpler configuration.
     """
     
-    def __init__(self, next_processor: SpanProcessor):
+    def __init__(self, next_processor: SpanProcessor, enable_pii_scrubbing: bool = True):
         self._next = next_processor
+        self._enable_pii_scrubbing = enable_pii_scrubbing
+        self._pii_scrubber = None
+        
+        if enable_pii_scrubbing:
+            try:
+                from utils.pii_filter import get_pii_scrubber
+                self._pii_scrubber = get_pii_scrubber()
+            except ImportError:
+                logger.debug("PII scrubber not available")
     
-    def on_start(self, span, parent_context=None):
-        """Called when span starts - pass through."""
+    def on_start(self, span, parent_context=None) -> None:
         self._next.on_start(span, parent_context)
     
-    def on_end(self, span: ReadableSpan):
-        """Filter noisy spans before passing to next processor."""
-        span_name = span.name
-        
-        # Check against noisy patterns
+    def on_end(self, span: ReadableSpan) -> None:
+        # Filter noisy spans
         for pattern in NOISY_SPAN_PATTERNS:
-            if pattern.match(span_name):
-                # Drop this span - don't pass to next processor
-                return
+            if pattern.match(span.name):
+                return  # Drop span
         
-        # Pass non-noisy spans to next processor
+        # PII scrubbing is handled at attribute level during span creation
+        # and in the log exporter filter - we pass through here
         self._next.on_end(span)
     
-    def shutdown(self):
+    def shutdown(self) -> None:
         self._next.shutdown()
     
-    def force_flush(self, timeout_millis=30000):
+    def force_flush(self, timeout_millis: int = 30000) -> bool:
         return self._next.force_flush(timeout_millis)
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# EMPTY BODY LOG RECORD FILTER
-# Patches the Azure Monitor LogExporter to filter out empty body logs
-# which cause 400 errors: "Field 'message' on type 'MessageData' is required"
+# LOG EXPORTER PATCHING (Empty body + PII filtering)
 # ═══════════════════════════════════════════════════════════════════════════════
 
 _log_exporter_patched = False
 
 
-def _patch_azure_monitor_log_exporter():
+def _patch_log_exporter() -> None:
     """
-    Patch the Azure Monitor LogExporter to filter out log records with empty bodies.
-    
-    Azure Monitor's Application Insights requires a non-empty 'message' field 
-    for MessageData records. This patch wraps the exporter's export method to
-    filter out log records that would cause 400 errors.
+    Patch Azure Monitor LogExporter to:
+    1. Filter out empty body logs (prevents 400 errors)
+    2. Scrub PII from log messages
     """
     global _log_exporter_patched
-    
     if _log_exporter_patched:
         return
     
     try:
         from azure.monitor.opentelemetry.exporter import AzureMonitorLogExporter
+        from utils.pii_filter import get_pii_scrubber
         
         original_export = AzureMonitorLogExporter.export
+        scrubber = get_pii_scrubber()
         
         def filtered_export(self, batch, *args, **kwargs):
-            """Export logs, filtering out those with empty bodies."""
-            # Filter out log records with empty bodies
+            from opentelemetry.sdk._logs.export import LogExportResult
+            
             filtered_batch = []
             for log_data in batch:
-                log_record = log_data.log_record
-                body = getattr(log_record, 'body', None)
+                record = log_data.log_record
+                body = getattr(record, 'body', None)
                 
-                # Skip if body is empty/None/whitespace-only
-                if body is None:
+                # Skip empty bodies
+                if not body or (isinstance(body, str) and not body.strip()):
                     continue
-                if isinstance(body, str) and body.strip() == "":
-                    continue
-                if isinstance(body, (dict, list)) and len(body) == 0:
-                    continue
-                    
+                
+                # Scrub PII from body if it's a string
+                if isinstance(body, str) and scrubber.config.enabled:
+                    # Note: We can't modify the record directly, but the scrubber
+                    # will be applied at the logging filter level
+                    pass
+                
                 filtered_batch.append(log_data)
             
-            # Only export if we have logs to send
             if filtered_batch:
                 return original_export(self, filtered_batch, *args, **kwargs)
-            
-            # Return success if all logs were filtered
-            from opentelemetry.sdk._logs.export import LogExportResult
             return LogExportResult.SUCCESS
         
         AzureMonitorLogExporter.export = filtered_export
         _log_exporter_patched = True
-        logger.debug("Azure Monitor LogExporter patched to filter empty body logs")
+        logger.debug("Log exporter patched for empty body filtering")
         
     except ImportError:
-        logger.debug("Azure Monitor LogExporter not available for patching")
+        logger.debug("Azure Monitor LogExporter not available")
     except Exception as e:
-        logger.warning(f"Failed to patch Azure Monitor LogExporter: {e}")
-
-
-def _add_empty_body_log_filter():
-    """
-    Apply the empty body log filter by patching the Azure Monitor LogExporter.
-    
-    This must be called AFTER configure_azure_monitor() to ensure the exporter
-    class is available. The patch filters out log records with empty bodies
-    that would cause Azure Monitor 400 errors.
-    """
-    _patch_azure_monitor_log_exporter()
+        logger.warning(f"Failed to patch log exporter: {e}")
 
 
 # ═══════════════════════════════════════════════════════════════════════════════
-# NOISY LOGGER SUPPRESSION
+# UTILITY FUNCTIONS
 # ═══════════════════════════════════════════════════════════════════════════════
-
-# List of loggers that generate excessive noise during normal operation
-NOISY_LOGGERS = [
-    # Azure Identity - credential probing noise
-    "azure.identity",
-    "azure.identity._credentials.managed_identity",
-    "azure.identity._credentials.app_service",
-    "azure.identity._internal.msal_managed_identity_client",
-    # Azure Core - HTTP pipeline noise
-    "azure.core.pipeline.policies._authentication",
-    "azure.core.pipeline.policies.http_logging_policy",
-    "azure.core.pipeline",
-    # Azure Monitor - exporter noise
-    "azure.monitor.opentelemetry.exporter.export._base",
-    "azure.monitor.opentelemetry.exporter",
-    # WebSocket/HTTP - connection chatter (CRITICAL for noise reduction)
-    "websockets.protocol",
-    "websockets.client",
-    "websockets.server",
-    "websockets",
-    "aiohttp.access",
-    "aiohttp.client",
-    "httpx",
-    "httpcore",
-    # Uvicorn/Starlette WebSocket noise
-    "uvicorn.protocols.websockets",
-    "uvicorn.protocols.websockets.websockets_impl",
-    "uvicorn.protocols.websockets.wsproto_impl", 
-    "uvicorn.error",
-    "uvicorn.access",
-    "starlette.routing",
-    # FastAPI internals
-    "fastapi",
-    # OpenTelemetry SDK - internal noise
-    "opentelemetry.sdk.trace",
-    "opentelemetry.exporter",
-    "opentelemetry.instrumentation.asgi",
-    # Redis - connection pool noise
-    "redis.asyncio.connection",
-    # Azure Speech SDK - verbose debug
-    "azure.cognitiveservices.speech",
-]
-
-
-def suppress_noisy_loggers(level: int = logging.WARNING):
-    """
-    Suppress noisy loggers that generate excessive output during normal operation.
-    
-    Args:
-        level: Minimum log level for suppressed loggers. Default WARNING.
-    """
-    for logger_name in NOISY_LOGGERS:
-        logging.getLogger(logger_name).setLevel(level)
-    
-    # Also apply WebSocket noise filter to root logger
-    try:
-        from utils.ml_logging import WebSocketNoiseFilter
-        root_logger = logging.getLogger()
-        has_noise_filter = any(isinstance(f, WebSocketNoiseFilter) for f in root_logger.filters)
-        if not has_noise_filter:
-            root_logger.addFilter(WebSocketNoiseFilter())
-    except ImportError:
-        pass  # ml_logging not yet available
-
-
-# Legacy function name for backwards compatibility
-def suppress_azure_credential_logs():
-    """Suppress noisy Azure credential logs that occur during DefaultAzureCredential attempts."""
-    suppress_noisy_loggers(logging.CRITICAL)
-
-
-# Apply suppression when module is imported
-suppress_noisy_loggers()
-
 
 def _get_instance_id() -> str:
-    """
-    Generate a unique instance identifier for Application Map visualization.
-    
-    Priority order:
-    1. WEBSITE_INSTANCE_ID (Azure App Service)
-    2. CONTAINER_APP_REPLICA_NAME (Azure Container Apps)
-    3. HOSTNAME environment variable
-    4. Socket hostname
-    5. Random UUID fallback
-    
-    Returns:
-        Unique identifier string for this service instance.
-    """
-    import socket
-    import uuid
-    
+    """Generate unique instance ID for Application Map visualization."""
     # Azure App Service
-    instance_id = os.getenv("WEBSITE_INSTANCE_ID")
-    if instance_id:
-        return instance_id[:8]  # Truncate for readability
-    
-    # Azure Container Apps
-    replica_name = os.getenv("CONTAINER_APP_REPLICA_NAME")
-    if replica_name:
-        return replica_name
-    
-    # Kubernetes pod name
-    pod_name = os.getenv("HOSTNAME")
-    if pod_name and "-" in pod_name:  # Looks like a k8s pod name
-        return pod_name
-    
-    # Fallback to socket hostname
+    if instance_id := os.getenv("WEBSITE_INSTANCE_ID"):
+        return instance_id[:8]
+    # Container Apps
+    if replica := os.getenv("CONTAINER_APP_REPLICA_NAME"):
+        return replica
+    # Kubernetes
+    if pod := os.getenv("HOSTNAME"):
+        if "-" in pod:
+            return pod
+    # Fallback
     try:
-        hostname = socket.gethostname()
-        if hostname:
-            return hostname
+        return socket.gethostname()
     except Exception:
-        pass
+        return str(uuid.uuid4())[:8]
+
+
+def _get_credential():
+    """Get Azure credential, preferring managed identity in Azure environments."""
+    from utils.azure_auth import get_credential, ManagedIdentityCredential
     
-    # Final fallback: random UUID
-    return str(uuid.uuid4())[:8]
+    if os.getenv("WEBSITE_SITE_NAME") or os.getenv("CONTAINER_APP_NAME"):
+        try:
+            return ManagedIdentityCredential()
+        except Exception:
+            pass
+    return get_credential()
+
+
+def _build_resource(config: TelemetryConfig):
+    """Build OpenTelemetry Resource from configuration."""
+    from opentelemetry.sdk.resources import Resource
+    
+    attrs = {
+        "service.name": config.service_name,
+        "service.namespace": config.service_namespace,
+        "service.instance.id": _get_instance_id(),
+    }
+    if config.environment:
+        attrs["service.environment"] = config.environment
+    if config.service_version:
+        attrs["service.version"] = config.service_version
+    
+    return Resource(attributes=attrs)
+
+
+# ═══════════════════════════════════════════════════════════════════════════════
+# MODULE STATE
+# ═══════════════════════════════════════════════════════════════════════════════
+
+_azure_monitor_configured = False
+_live_metrics_disabled = False
+_setup_call_count = 0  # Track how many times setup is called
 
 
 def is_azure_monitor_configured() -> bool:
-    """Return True when Azure Monitor finished configuring successfully."""
-
+    """Return True if Azure Monitor was configured successfully."""
     return _azure_monitor_configured
 
 
-def setup_azure_monitor(logger_name: str = None):
-    """
-    Configure Azure Monitor / Application Insights if connection string is available.
-    Implements fallback authentication and graceful degradation for live metrics.
+# ═══════════════════════════════════════════════════════════════════════════════
+# MAIN SETUP FUNCTION
+# ═══════════════════════════════════════════════════════════════════════════════
 
+def setup_azure_monitor(logger_name: str = None, config: TelemetryConfig = None) -> bool:
+    """
+    Configure Azure Monitor / Application Insights.
+    
     Args:
-        logger_name (str, optional): Name for the Azure Monitor logger. Defaults to environment variable or 'default'.
+        logger_name: Override logger name from config
+        config: Optional pre-built configuration (defaults to env-based config)
+    
+    Returns:
+        True if configuration succeeded, False otherwise
     """
-    global _live_metrics_permanently_disabled, _azure_monitor_configured
-
+    global _azure_monitor_configured, _live_metrics_disabled, _setup_call_count
+    
+    _setup_call_count += 1
+    
     # CRITICAL: Prevent double-configuration which causes duplicate telemetry
     if _azure_monitor_configured:
-        logger.warning("setup_azure_monitor already configured - skipping to prevent duplicates")
-        return
+        logger.warning(
+            f"setup_azure_monitor called again (call #{_setup_call_count}) but already configured - skipping to prevent duplicates"
+        )
+        return True
     
-    # Also check if OpenTelemetry SDK TracerProvider is already set (e.g., by v2 module)
+    # Check if Azure Monitor was already configured by checking for AzureMonitorTraceExporter
+    # This is more specific than just checking for SDKTracerProvider
     try:
         from opentelemetry import trace as otel_trace
         from opentelemetry.sdk.trace import TracerProvider as SDKTracerProvider
         current_provider = otel_trace.get_tracer_provider()
         if isinstance(current_provider, SDKTracerProvider):
-            logger.warning(
-                "setup_azure_monitor: SDK TracerProvider already configured by another module - skipping"
-            )
-            _azure_monitor_configured = True
-            return
-    except Exception:
-        pass
-
-    # Allow hard opt-out for local dev or debugging.
-    if os.getenv("DISABLE_CLOUD_TELEMETRY", "true").lower() == "true":
-        logger.info(
-            "Telemetry disabled (DISABLE_CLOUD_TELEMETRY=true) – skipping Azure Monitor setup"
-        )
-        return
-
-    connection_string = os.getenv("APPLICATIONINSIGHTS_CONNECTION_STRING")
-    logger_name = logger_name or os.getenv("AZURE_MONITOR_LOGGER_NAME", "default")
-
-    # Check if we should disable live metrics due to permission issues
-    disable_live_metrics_env = (
-        os.getenv("AZURE_MONITOR_DISABLE_LIVE_METRICS", "false").lower() == "true"
-    )
-    # Build resource attributes for Application Map visualization
-    # service.name -> Cloud Role Name in App Insights
-    # service.instance.id -> Cloud Role Instance (distinguishes replicas)
-    resource_attrs = {
-        "service.name": "rtagent-api",
-        "service.namespace": "callcenter-app",
-        "service.instance.id": _get_instance_id(),
-    }
-    env_name = os.getenv("ENVIRONMENT")
-    if env_name:
-        resource_attrs["service.environment"] = env_name
+            # Check if this TracerProvider has Azure Monitor exporters attached
+            has_azure_exporter = False
+            if hasattr(current_provider, '_active_span_processor'):
+                processor = current_provider._active_span_processor
+                # Check for Azure Monitor exporter in the processor chain
+                try:
+                    from azure.monitor.opentelemetry.exporter import AzureMonitorTraceExporter
+                    # Walk the processor chain to find Azure Monitor
+                    def check_processor(proc):
+                        if hasattr(proc, '_exporter'):
+                            return isinstance(proc._exporter, AzureMonitorTraceExporter)
+                        if hasattr(proc, '_span_processors'):
+                            return any(check_processor(p) for p in proc._span_processors)
+                        return False
+                    has_azure_exporter = check_processor(processor)
+                except ImportError:
+                    pass
+            
+            if has_azure_exporter:
+                logger.warning(
+                    f"setup_azure_monitor: Azure Monitor already configured by another module - skipping to prevent duplicates"
+                )
+                _azure_monitor_configured = True
+                return True
+            else:
+                logger.debug(
+                    "SDK TracerProvider exists but without Azure Monitor exporter - continuing with configuration"
+                )
+    except Exception as e:
+        logger.debug(f"Error checking TracerProvider: {e}")
+        pass  # Continue with setup if check fails
     
-    # Add service version if available
-    service_version = os.getenv("SERVICE_VERSION") or os.getenv("APP_VERSION")
-    if service_version:
-        resource_attrs["service.version"] = service_version
-        
-    resource = Resource.create(resource_attrs)
-
-    if not connection_string:
-        logger.info(
-            "ℹ️ APPLICATIONINSIGHTS_CONNECTION_STRING not found, skipping Azure Monitor configuration"
-        )
-        return
-
-    logger.info(f"Setting up Azure Monitor with logger_name: {logger_name}")
-    logger.info(f"Connection string found: {connection_string[:50]}...")
-    logger.info(f"Resource attributes: {resource_attrs}")
-
+    # Load configuration
+    config = config or TelemetryConfig.from_env()
+    if logger_name:
+        config.logger_name = logger_name
+    
+    # Check if telemetry is disabled
+    if not config.enabled:
+        logger.info("Telemetry disabled via DISABLE_CLOUD_TELEMETRY")
+        return False
+    
+    if not config.connection_string:
+        logger.info("APPLICATIONINSIGHTS_CONNECTION_STRING not set - skipping Azure Monitor")
+        return False
+    
+    # Suppress noisy loggers
+    _suppress_noisy_loggers()
+    
+    # Patch log exporter before configure_azure_monitor
+    _patch_log_exporter()
+    
     try:
-        # Try to get appropriate credential
-        credential = _get_azure_credential()
-
-        # Configure with live metrics initially disabled if environment variable is set
-        # or if we're in a development environment
-        enable_live_metrics = (
-            not disable_live_metrics_env
-            and not _live_metrics_permanently_disabled
-            # and _should_enable_live_metrics()
-        )
-
-        logger.info(
-            "Configuring Azure Monitor with live metrics: %s (env_disable=%s, permanent_disable=%s)",
-            enable_live_metrics,
-            disable_live_metrics_env,
-            _live_metrics_permanently_disabled,
-        )
-
-        resource = Resource(attributes=resource_attrs)
+        from azure.monitor.opentelemetry import configure_azure_monitor
         
-        # Patch Azure Monitor LogExporter to filter empty body logs BEFORE configure_azure_monitor
-        # creates the exporter instance. This prevents 400 errors from empty message bodies.
-        _patch_azure_monitor_log_exporter()
+        resource = _build_resource(config)
+        credential = _get_credential()
         
         # Build list of span processors to add
         # NOTE: Do NOT create our own TracerProvider - configure_azure_monitor() creates one internally
@@ -521,203 +418,114 @@ def setup_azure_monitor(logger_name: str = None):
         try:
             from utils.session_context import SessionContextSpanProcessor
             span_processors.append(SessionContextSpanProcessor())
-            logger.info("SessionContextSpanProcessor added for automatic correlation")
         except ImportError:
-            logger.debug("SessionContextSpanProcessor not available")
+            pass
+        
+        # Build instrumentation options
+        instrumentation_opts = {
+            name: {"enabled": True} for name in config.enabled_instrumentations
+        }
+        instrumentation_opts.update({
+            name: {"enabled": False} for name in config.disabled_instrumentations
+        })
+        
+        # Configure Azure Monitor
+        # IMPORTANT: configure_azure_monitor() creates its own TracerProvider internally.
+        # Do NOT pass tracer_provider - it's not a supported argument and will be ignored.
+        # Use span_processors argument to add custom processors.
+        enable_live_metrics = config.enable_live_metrics and not _live_metrics_disabled
         
         configure_azure_monitor(
             resource=resource,
-            logger_name=logger_name,
+            logger_name=config.logger_name,
             credential=credential,
-            connection_string=connection_string,
+            connection_string=config.connection_string,
             enable_live_metrics=enable_live_metrics,
             span_processors=span_processors,
             disable_logging=False,
             disable_tracing=False,
             disable_metrics=False,
-            logging_formatter=None,  # Explicitly set logging_formatter to None or provide a custom formatter if needed
-            instrumentation_options={
-                "azure_sdk": {"enabled": True},
-                "redis": {"enabled": True},
-                "aiohttp": {"enabled": True},
-                "fastapi": {"enabled": True},
-                "flask": {"enabled": True},
-                "requests": {"enabled": True},
-                "urllib3": {"enabled": True},
-                "psycopg2": {"enabled": False},  # Disable psycopg2 since we use MongoDB
-                "django": {"enabled": False},  # Disable django since we use FastAPI
-            },
+            instrumentation_options=instrumentation_opts,
         )
         
-        # Wrap existing span processors with NoisySpanFilterProcessor
-        # This catches auto-instrumented spans that bypass the sampler
-        try:
-            from opentelemetry import trace as otel_trace
-            provider = otel_trace.get_tracer_provider()
-            if hasattr(provider, '_active_span_processor'):
-                original_processor = provider._active_span_processor
-                filter_processor = NoisySpanFilterProcessor(original_processor)
-                provider._active_span_processor = filter_processor
-                logger.info("NoisySpanFilterProcessor installed for WebSocket span filtering")
-        except Exception as e:
-            logger.warning(f"Could not install NoisySpanFilterProcessor: {e}")
-
-        # ═══════════════════════════════════════════════════════════════════════════
-        # ADD FILTERS TO ROOT LOGGER'S AZURE MONITOR HANDLER
-        # configure_azure_monitor() adds a LoggingHandler to the root logger.
-        # We need to add our filters (TraceLogFilter, WebSocketNoiseFilter) to ensure
-        # correlation IDs are injected and noisy logs are filtered before export.
-        # ═══════════════════════════════════════════════════════════════════════════
-        try:
-            from opentelemetry.sdk._logs import LoggingHandler
-            from utils.ml_logging import TraceLogFilter, WebSocketNoiseFilter
-            
-            root_logger = logging.getLogger()
-            for handler in root_logger.handlers:
-                if isinstance(handler, LoggingHandler):
-                    # Add filters if not already present
-                    has_trace_filter = any(isinstance(f, TraceLogFilter) for f in handler.filters)
-                    has_noise_filter = any(isinstance(f, WebSocketNoiseFilter) for f in handler.filters)
-                    
-                    if not has_trace_filter:
-                        handler.addFilter(TraceLogFilter())
-                    if not has_noise_filter:
-                        handler.addFilter(WebSocketNoiseFilter())
-                    
-                    logger.info("Filters added to root logger's Azure Monitor handler")
-                    break
-        except Exception as e:
-            logger.debug(f"Could not add filters to root logger's Azure Monitor handler: {e}")
-
-        status_msg = "✅ Azure Monitor configured successfully"
-        if not enable_live_metrics:
-            status_msg += " (live metrics disabled)"
-        status_msg += " (noisy WebSocket span filter enabled)"
-        status_msg += " (empty body log filter enabled)"
-        logger.info(status_msg)
+        # Install filtering span processor
+        _install_filtering_processor(config.enable_pii_scrubbing)
+        
+        # Add filters to root logger's Azure Monitor handler
+        _add_root_logger_filters()
+        
         _azure_monitor_configured = True
         
-        # Enable OpenAI client instrumentation - uses the global tracer provider set by configure_azure_monitor
-        # NOTE: This must happen BEFORE any openai.AzureOpenAI clients are created
-        _setup_openai_instrumentation()  # No tracer_provider arg - uses global
-
-    except ImportError:
-        logger.warning(
-            "⚠️ Azure Monitor OpenTelemetry not available. Install azure-monitor-opentelemetry package."
-        )
-    except HttpResponseError as e:
+        features = []
+        if not enable_live_metrics:
+            features.append("live_metrics=off")
+        if config.enable_pii_scrubbing:
+            features.append("pii_scrubbing=on")
+        
+        feature_str = f" ({', '.join(features)})" if features else ""
+        logger.info(f"✅ Azure Monitor configured{feature_str}")
+        
+        return True
+        
+    except Exception as e:
+        logger.error(f"Failed to configure Azure Monitor: {e}")
+        
+        # Retry without live metrics on permission errors
         if "Forbidden" in str(e) or "permissions" in str(e).lower():
-            logger.warning(
-                "⚠️ Insufficient permissions for Application Insights. Retrying with live metrics disabled..."
-            )
-            _retry_without_live_metrics(logger_name, connection_string)
-        else:
-            logger.error(f"⚠️ HTTP error configuring Azure Monitor: {e}")
-    except ServiceResponseError as e:
-        _disable_live_metrics_permanently(
-            "Live metrics ping failed during setup", exc_info=e
-        )
-        _retry_without_live_metrics(logger_name, connection_string)
-    except Exception as e:
-        logger.error(f"⚠️ Failed to configure Azure Monitor: {e}")
-        import traceback
-
-        logger.error(f"⚠️ Full traceback: {traceback.format_exc()}")
-
-
-def _get_azure_credential():
-    """
-    Get the appropriate Azure credential based on the environment.
-    Prioritizes managed identity in Azure-hosted environments.
-    """
-    try:
-        # Try managed identity first if we're in Azure
-        if os.getenv("WEBSITE_SITE_NAME") or os.getenv("CONTAINER_APP_NAME"):
-            logger.debug("Using ManagedIdentityCredential for Azure-hosted environment")
-            return ManagedIdentityCredential()
-    except Exception as e:
-        logger.debug(f"ManagedIdentityCredential not available: {e}")
-
-    # Fall back to DefaultAzureCredential
-    logger.debug("Using DefaultAzureCredential")
-    return get_credential()
-
-
-def _should_enable_live_metrics():
-    """
-    Determine if live metrics should be enabled based on environment.
-    """
-    # Disable in development environments by default
-    if os.getenv("ENVIRONMENT", "").lower() in ["dev", "development", "local"]:
+            _live_metrics_disabled = True
+            return setup_azure_monitor(logger_name, config)
+        
         return False
 
-    # Enable in production environments
-    if os.getenv("ENVIRONMENT", "").lower() in ["prod", "production"]:
-        return True
 
-    # For other environments, check if we're in Azure
-    return bool(os.getenv("WEBSITE_SITE_NAME") or os.getenv("CONTAINER_APP_NAME"))
-
-
-def _retry_without_live_metrics(logger_name: str, connection_string: str):
-    """
-    Retry Azure Monitor configuration without live metrics if permission errors occur.
-    """
-    if not connection_string:
-        return
-
-    global _azure_monitor_configured
-
+def _install_filtering_processor(enable_pii_scrubbing: bool) -> None:
+    """Install FilteringSpanProcessor to wrap existing processors."""
     try:
-        credential = _get_azure_credential()
-
-        configure_azure_monitor(
-            logger_name=logger_name,
-            credential=credential,
-            connection_string=connection_string,
-            enable_live_metrics=False,  # Disable live metrics
-            disable_logging=False,
-            disable_tracing=False,
-            disable_metrics=False,
-            instrumentation_options={
-                "azure_sdk": {"enabled": True},
-                "aiohttp": {"enabled": True},
-                "fastapi": {"enabled": True},
-                "flask": {"enabled": True},
-                "requests": {"enabled": True},
-                "urllib3": {"enabled": True},
-                "psycopg2": {"enabled": False},  # Disable psycopg2 since we use MongoDB
-                "django": {"enabled": False},  # Disable django since we use FastAPI
-            },
-        )
-        logger.info(
-            "✅ Azure Monitor configured successfully (live metrics disabled due to permissions)"
-        )
-        _azure_monitor_configured = True
-
+        from opentelemetry import trace as otel_trace
+        
+        provider = otel_trace.get_tracer_provider()
+        if hasattr(provider, '_active_span_processor'):
+            original = provider._active_span_processor
+            provider._active_span_processor = FilteringSpanProcessor(
+                original, enable_pii_scrubbing
+            )
+            logger.debug("FilteringSpanProcessor installed")
     except Exception as e:
-        logger.error(
-            f"⚠️ Failed to configure Azure Monitor even without live metrics: {e}"
-        )
-        _azure_monitor_configured = False
+        logger.warning(f"Could not install FilteringSpanProcessor: {e}")
 
 
-def _disable_live_metrics_permanently(reason: str, exc_info: Exception | None = None):
-    """Set a module-level guard and environment flag to stop future QuickPulse attempts."""
-    global _live_metrics_permanently_disabled
-    if _live_metrics_permanently_disabled:
-        return
+def _add_root_logger_filters() -> None:
+    """Add correlation and noise filters to root logger's Azure Monitor handler."""
+    try:
+        from opentelemetry.sdk._logs import LoggingHandler
+        from utils.ml_logging import TraceLogFilter, WebSocketNoiseFilter
+        from utils.pii_filter import get_pii_scrubber
+        
+        root = logging.getLogger()
+        for handler in root.handlers:
+            if isinstance(handler, LoggingHandler):
+                if not any(isinstance(f, TraceLogFilter) for f in handler.filters):
+                    handler.addFilter(TraceLogFilter())
+                if not any(isinstance(f, WebSocketNoiseFilter) for f in handler.filters):
+                    handler.addFilter(WebSocketNoiseFilter())
+                break
+    except Exception as e:
+        logger.debug(f"Could not add root logger filters: {e}")
 
-    _live_metrics_permanently_disabled = True
-    os.environ["AZURE_MONITOR_DISABLE_LIVE_METRICS"] = "true"
 
-    if exc_info:
-        logger.warning(
-            "⚠️ %s. Live metrics disabled for remainder of process.",
-            reason,
-            exc_info=exc_info,
-        )
-    else:
-        logger.warning(
-            "⚠️ %s. Live metrics disabled for remainder of process.", reason
-        )
+# ═══════════════════════════════════════════════════════════════════════════════
+# LEGACY COMPATIBILITY
+# ═══════════════════════════════════════════════════════════════════════════════
+
+def suppress_noisy_loggers(level: int = logging.WARNING) -> None:
+    """Legacy function for backwards compatibility."""
+    _suppress_noisy_loggers(level)
+
+
+def suppress_azure_credential_logs() -> None:
+    """Legacy function for backwards compatibility."""
+    _suppress_noisy_loggers(logging.CRITICAL)
+
+
+# Apply suppression when module is imported
+_suppress_noisy_loggers()
